AutoGUILD & AutoCHANNEL parameters
