+++ Examples/AutomaticGeneration/main_autochannel.py
@@ -15,7 +15,7 @@ from daf.logging.logger_json import LoggerJSON
 from daf.messagedata.textdata import TextMessageData
 from daf.message.text_based import TextMESSAGE
 from daf.client import ACCOUNT
-from daf.dtypes import FILE
+from daf.messagedata import FILE
 from daf.message.messageperiod import FixedDurationPeriod
 from datetime import timedelta
 from daf.guild.guilduser import GUILD
