+++ Examples/Logging/CSV/main_csv_logging.py
@@ -15,7 +15,7 @@ from daf.logging.logger_csv import LoggerCSV
 from daf.messagedata.textdata import TextMessageData
 from daf.message.text_based import TextMESSAGE
 from daf.client import ACCOUNT
-from daf.dtypes import FILE
+from daf.messagedata import FILE import FILE
 from daf.message.messageperiod import FixedDurationPeriod
 from datetime import timedelta
 from daf.guild.guilduser import GUILD
