+++ docs/source/guide/index.rst
@@ -5,7 +5,7 @@ Guide
 
 The following pages contain guides to running the Discord Advertisement Framework in 2 main modes:
 
-1. :ref:`Core mode (console / programming) <Quickstart (core)>`
+1. :ref:`Core mode (console / programming) <Framework control (core)>`
 2. :ref:`Graphical mode (GUI) <Quickstart (GUI)>`
 
 .. toctree::
