+++ src/daf/responder/guildresponder.py
@@ -1,14 +1,13 @@
 from typing import List
 from typeguard import typechecked
 
+from .constraints import BaseGuildConstraint
+from ..misc.instance_track import track_id
+from ..misc.doc import doc_category
+from .actions import BaseResponse
 from .base import ResponderBase
 from ..logic import BaseLogic
-from .constraints import BaseGuildConstraint
-from .actions import BaseResponse
 from ..events import EventID
-from ..misc.doc import doc_category
-from ..misc.instance_track import track_id
-
 
 import asyncio_event_hub as aeh
 import _discord as discord
